chore: add translation and device as arg to renderError
