+++ apps/ledger-live-desktop/src/renderer/components/DeviceAction/index.tsx
@@ -241,6 +241,7 @@ export const DeviceActionDefaultRendering = <R, H, P>({
 
   if (inWrongDeviceForAccount) {
     return renderInWrongAppForAccount({
+      t,
       onRetry,
       accountName: inWrongDeviceForAccount.accountName,
     });
@@ -253,6 +254,7 @@ export const DeviceActionDefaultRendering = <R, H, P>({
       error instanceof UpdateYourApp
     ) {
       return renderError({
+        t,
         error,
         managerAppName: error.managerAppName,
       });
@@ -260,6 +262,7 @@ export const DeviceActionDefaultRendering = <R, H, P>({
 
     if (error instanceof LatestFirmwareVersionRequired) {
       return renderError({
+        t,
         error,
         requireFirmwareUpdate: true,
       });
@@ -272,6 +275,7 @@ export const DeviceActionDefaultRendering = <R, H, P>({
       (error instanceof TransportStatusError && error.message.includes("0x6d06"))
     ) {
       return renderError({
+        t,
         error: new DeviceNotOnboarded(),
         withOnboardingCTA: true,
         info: true,
@@ -280,6 +284,7 @@ export const DeviceActionDefaultRendering = <R, H, P>({
 
     if (error instanceof NoSuchAppOnProvider) {
       return renderError({
+        t,
         error,
         withOpenManager: true,
         withExportLogs: true,
@@ -287,9 +292,11 @@ export const DeviceActionDefaultRendering = <R, H, P>({
     }
 
     return renderError({
+      t,
       error,
       onRetry,
       withExportLogs: true,
+      device: device ?? undefined,
     });
   }
