+++ apps/ledger-live-desktop/src/renderer/components/ErrorDisplay.tsx
@@ -1,0 +1,26 @@
+import { useTranslation } from "react-i18next";
+import { renderError } from "~/renderer/components/DeviceAction/rendering";
+
+export type ErrorDisplayProps = {
+  error: Error;
+  onRetry?: () => void;
+  withExportLogs?: boolean;
+  list?: boolean;
+  supportLink?: string;
+  warning?: boolean;
+};
+
+const ErrorDisplay = ({
+  error,
+  onRetry,
+  withExportLogs,
+  list,
+  supportLink,
+  warning,
+}: ErrorDisplayProps) => {
+  const { t } = useTranslation();
+
+  renderError({ t, error, onRetry, withExportLogs, list, supportLink, warning });
+};
+
+export default ErrorDisplay;
